requirements.txt
torch
transformers
earthengine-api
python-dotenv
numpy
matplotlib
